Prepare library for getBlob
Completed implementation of getBlob. The documentation
for it as well as BlobQuery and sendRequest(_:) is roughly
complete, save for some spelling and grammar checks. A
pdsURL parameter was also added since getBlob is a static
method. The queryItems were added and both it and the
createRequest method were moved to the do catch block.