Add implementation for getSyncRecord
It appears that this method is similar to getRepoRecord.
However, it hasn't been tested yet. The API playground
Bluesky has doesn't seem to work, but it's been speculated
that it's not redirecting to bsky.social. This will therfore need
to be tested on another server connected to AT Protocol.

For now, this will be dormant until it can be tested and a
full understanding of .cat files (including how to parse
them) are made.