Refactor sendRequest methods
Three of the sendRequest methods contain duplicate
code. As a result, it violated the DRY pattern. This
commit fixes this. An unrelated empty line
deletion has also been performed in this commit.

There are a couple of other methods that do similar things,
but they'll be left alone for now.