Add implementation of getProfile
According to the Bluesky docs, this can be called with or without
authorization. However, when attempting to call this method
without authorization, the errors indicate that authorization is
required. Bluesky will need to clarify their position on this API call.
If it's determined that it really isn't required, then the following
should happen with this method:
- The method will be static, as all ATProtoKit methods that
deal with API calls should be static.
- "accessToken" and "pdsURL" should be arguments.
- "session.pdsURL" will be simply "pdsURL".
- The commented-out documentation that speaks about passing
"accessToken" for authorization should be added.
The commented-out code should be uncommented; the line
on hold should be removed.

If it's determined that authorization is required after all, then the
commented-out documentation and code needs to be removed.

Also, ActorGetProfileQuery now has a public init in order to
access it.