Change getProfile to unauthenticated method
After asking around, it turns out that Bluesky's docs were
indeed accurate. Therefore, the following changes have
been made:
- Changed the method to be static.
- Added the "accessToken" and "pdsURL" arguments.
- Moved the documentation that speaks about the accessToken
to the appropiate spot, and added a note about using a pdsURL
that's used for unauthenticated API calls if needed.
- Created a static method for determining the pdsURL and replaced
the part of the guard statement that unwraps the the URL.
- Removed and uncommented the appropiate code.

Seaprately, a bullet point was added to README that asks contributors
to make methods within the ATProtoKit class static if authentication
is either required or optional.