Change Note tags to Important tags
This is needed to grab more attention to how the
library handles items that go over the maxiumum
limit. If it sees that something is going over the limit,
it will truncate the length until it's at its maximum
length. This is to help prevent unexpected crashes.
However, some may not realize that this is happening.
This change will help to put things into perspective.