Add getSuggestions and fix issue
The lexicon and method have been implemented
and tested without much of an issue.

As for the issue, this is where the "without much of
an issue" comes into play. There seems to be an issue
wherre the library is unable to handle dates that are on
1970, or contains fracional seconds. Because of this, the
library will refuse to further decode the array. The solution
was to set the formatOptions of ISO8601DateFormatter to
use the Internet date time and fractional seconds. According
to Apple's documentation, however, this may have an issue
with performance, and I'd like to make sure performance hits
are kept to an absolute minimum. The singleton pattern in the
date formatter was supposed to be used as a temporary
measure. However, given how it appears to have managed to
prevent any form huge performance issues, it seems little choice
can be done here. The Singleton pattern is best to be avoided
in this library, but given how no other solution can be found at
this time, an exception will be made here.