Add SearchUsers and add docs
- Added the lexicon and implementation for searchUsers
and ActorSearchActors.
- Added documentation to ActorGetSuggestionsOutput.actors.

There's currently a bug in searchUsers: in spite the AT Protocol
documentation saying that authentication is optional, the API call
responds with an "Authentication Required" error. Until Bluesky
clarifies, this option will remain as-is. If it's understood that it is
required, then it will no longer become a static method.