Force deleteSession to use correct overload
There was a warning, saying that the variable is unused. This is in spite the fact
the overload used lacks a return type. "_ =" was used to no avail. The solution was
to explicitly expose "withEncodingBody" and setting it to nil so that the complier
is aware of which overload to use.