Add GetPostThread and fix decoding bug
There was an issue where there was a decoding mismatch
in the response. It turns out that PostUnion was erroneously
working in multiple objects. This meant that there was no
ThreadPostView object. There was supposed to be a
ThreadPostUnion as well, but due to user error, there wasn't.
The union enum was added and everything works again.