Make session optional and update methods
A decision was made to make session optional again to
allow methods that optionally or don't use authentication
to use the classes easily. This means making all of the
methods that are in a class that conforms to
ATProtoKitConfiguration have a check if there's an active
session (only applies to those that require or optionally
use authentication). The com.atproto.admin.*-related
methods have been changed in this commit, with more
to follow.

retrieveDID has also been changed to resolveHandle in
order to better align with the lexicon name.