Conform moderation methods to guidelines
The following changes were made:
- A note of the lexicon’s official description is added.
- The link and namespaced identifier of the lexicon is added.
- Expanded “NSID” to “Namespaced Identifier (NSID)”.
- Made sure the Result field is filled.
- Remove the static keyword.
- For methods where authentication isn’t used, pdsURL is now an
optional and the default value is set to nil.
- For methods where authentication isn’t required, accessToken
is now shouldAuthenticate, which is of type Bool and is set to false.
- For methods where authentication isn’t used, there will be a
sessionURL variable that is set to pdsURL (if the parameter is not
nil), or session.pdsURL (if the parameter is nil).
- For methods where authentication isn’t required, an additional
accessToken variable is used, returning nil (if shouldAuthenticate
is false) or “Bearer \(session.accessToken)” (if it returns true).
- The guard statement now throws
ATRequestPrepareError.invalidRequestURL.
- All requestBody parameters will have their own separate line.
- queryURL will be declared outside of the do-catch blocks.
- APIClientService.sendRequest()’s parameters will have their own
separate line, unless it only has one.