Fix decoding issue with records
Embeds were unable to be decoded mainly because
it was unable to find the "$type" key. The reason is because
ATRecorTypeRegistry.recordRegistry didn't set the key properly.
The key couldn't be set properly because the "type" property
couldn't be accessed. In order to access the property, all of the
record structs, as well as ATRecordProtocol itself, needed to make
the "type" property "static."