Move uploadBlob into separate file
For more consistency over all lexicon method's code design,
APIClientService.uploadBlob will be put into a separate file.
It will still be in APIClientService, but as an extension. Because of
this change, uploadBlob is unable to gain access to mimeType.
For this reason, mimeType's access has gone from "private"
to "package."