Change flow of adding record structs
This creates a new protocol called ATRecordConfiguration,
which adds a new property and method that will ensure
the ATProtoKit class has completed the process of adding
records to the registry. While the likihood of a method
accessing the registry prematurely are fairly slim, this will
make things safer for the library. This should fix one of the
issues related to Swift 6 language mode.