Fix decoding issue
When getting a JSON object that contains a starter pack, it may
incorrectly be decoded as a labeler view. This is because
labelerView and starterPackViewBasic contains the same requirements
and the same names as the requirements, but starterPackViewBasic
contains an additional requirement. The labelerView is above the
starterPackViewBasic in terms of order for one of the union enums and
as a result, any starter packs that goes to the union type will end up
always being decoded as a labelerView incorrectly. The solution is to
switch up their positions. However, this isn't ideal since the order is
needed for a future generator to create an easy structure. More work
is needed to figure out how to resolve this issue.

This commit also conforms labelerView and starterPackViewBasic to
the Swift API Design Guidelines.