Tweak ATProtocolConfiguration and UserSession
createAccount(), authenticate(), getSession(), refreshSession(),
resumeSession(), and deleteSession() will now check if there are any
service items within the DIDDocument.service array. If there are none,
or if "atproto_pds" isn't in there, then the methods will throw an error
and will not continue, since it means the account isn't valid.