Replace semaphore with DispatchGroup
This makes UnknownType the one to wait until the registration
is complete. Once the registration is complete, UnknownType will
be free to begin encoding and decoding.

This is a better solution, but it's still not perfect. A better solution
might be to make APIClientService wait until the registration
is complete.