Fix token expiration check logic
The logic was accidentally flipped. Instead of checking
if the expiration of the token surpassed the current time,
it instead asked if the expiration was ahead of the current
time, causing a bug where the token would always be
changed.

Furthermore, the time interval of the expiration is purposely
ahead by 10 seconds in order to give a buffer, just in case the
token expires mid-entry.