Begin to remove shouldAuthenticate
Any methods where the output changes depending
on whether authentication was used or not will
no longer have this argument. Instead, if a developer
wants to output the non-authenticated output,
they can create an instance of ATProtoKit without
the use of a SessionConfiguration instance.
Likewise, if they want an output that relies on
authentication, they can pass in the SessionConfiguration
instance into the ATProtoKit instance.