Defer completing List and ListItem creation methods
Trying to delete the record immediately after creating
it results in a failure; the server will respond by saying
the record doesn't exist. A sleep timer of 0.5 seconds
has been added before completing the method, which
should give the server time to allow for accessing and
deleting the record after.