Remove @_exported
This was mistakenly understood as a way to import
FoundationNetworking on all Swift files. It is now
understood that it re-exports the library to other
projects on top of importing ATProtoKit, which is
not ideal.

Doing this causes the project to fail on all
non-Apple platforms, so this will take time to
implement on all relevant Swift files going forward.