Fix PreferenceUnion encoding to include $type field
The PreferenceUnion encode method was using singleValueContainer which caused the loss of the required $type field when serializing preferences. This resulted in API errors when calling putPreferences:

"Input/preferences/0 must be an object which includes the \"$type\" property"

Changed the implementation to use container(keyedBy:) and explicitly encode the $type field for each preference type, ensuring AT Protocol compliance for union variants.