Change requestURL to bsky.social in describeRepo and listRepo
This is a temporary measure as it works. A future update will
allow the use of a protocol that will automatically resolve the
DID or handle and grab the service endpoint for use as the
base URL of the requestURL. It will only fallback to
bsky.social as a last resort.