Decode "type" parameter conditionally
There's a rare issue where output JSON objects
don't use the "type" parameter, causing the
unions to fail to decode the objects. The lack of the
"type" parameter is an invalid output when unions are involved.

There's no way to resolve this within ATProtoKit. So to mitgate
this, unions will now conditionally decode for the "type" parameter.
If it finds one, it will decode as normal. If the value isn't what the
union expects, then it will decode it as a dictionary, as normal.
If it doesn't find a "$type" parameter, then it will decode as
a dictionary, and insert the value of "type" as "unknown."